Move spm calculation into counter.
